requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
Flask==2.1.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mpmath==1.2.1
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests
six==1.16.0
sympy==1.10.1
tornado==6.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
Werkzeug==2.1.1